system/brain/requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
PyYAML==5.3.1
requests==2.25.0
urllib3==1.26.2
yapf==0.30.0
